Final changes based on Mack's comments
